chore: Moved the test-helpers file to the tests/helpers folder
