+++ tests/migration/ember-addon/steps/analyze-addon/edge-case-folders-are-missing.test.js
@@ -1,5 +1,5 @@
 import { analyzeAddon } from '../../../../../src/migration/ember-addon/steps/analyze-addon.js';
-import { assert, loadFixture, test } from '../../../../test-helpers.js';
+import { assert, loadFixture, test } from '../../../../helpers/testing.js';
 
 test('migration | ember-addon | steps | analyze-addon > edge case (folders are missing)', function () {
   const options = {
