+++ tests/migration/ember-addon/steps/augment-options/javascript.test.js
@@ -1,5 +1,5 @@
 import { augmentOptions } from '../../../../../src/migration/ember-addon/steps/augment-options.js';
-import { assert, loadFixture, test } from '../../../../test-helpers.js';
+import { assert, loadFixture, test } from '../../../../helpers/testing.js';
 
 test('migration | ember-addon | steps | augment-options > JavaScript', function () {
   const options = {
