+++ tests/migration/ember-addon/steps/use-relative-paths/base-case.test.js
@@ -1,5 +1,9 @@
 import { useRelativePaths } from '../../../../../src/migration/ember-addon/steps/use-relative-paths.js';
-import { assertFixture, loadFixture, test } from '../../../../test-helpers.js';
+import {
+  assertFixture,
+  loadFixture,
+  test,
+} from '../../../../helpers/testing.js';
 
 const fileMapping = new Map([
   [
