+++ tests/utils/files/create-files.test.js
@@ -1,5 +1,5 @@
 import { createFiles } from '../../../src/utils/files.js';
-import { assertFixture, loadFixture, test } from '../../test-helpers.js';
+import { assertFixture, loadFixture, test } from '../../helpers/testing.js';
 
 test('utils | files > createFiles', function () {
   const options = {
