+++ tests/utils/files/move-files.test.js
@@ -1,5 +1,5 @@
 import { moveFiles } from '../../../src/utils/files.js';
-import { assertFixture, loadFixture, test } from '../../test-helpers.js';
+import { assertFixture, loadFixture, test } from '../../helpers/testing.js';
 
 test('utils | files > moveFiles', function () {
   const options = {
