+++ tests/utils/files/remove-files.test.js
@@ -1,5 +1,5 @@
 import { removeFiles } from '../../../src/utils/files.js';
-import { assertFixture, loadFixture, test } from '../../test-helpers.js';
+import { assertFixture, loadFixture, test } from '../../helpers/testing.js';
 
 test('utils | files > removeFiles', function () {
   const options = {
